WIP: Porting CAGE to MonkeyBars
I got tired of managing all of the UI stuff by hand, and the MonkeyBars
library is designed for exactly this kind of stuff.

In the process of porting, I've re-done the threading model a bit,
pushing most Ruby code onto the event thread and making the PointSource
polling synchronous.

Oh, and I moved things into packages, too.